Add a reference to lib/$(Configuration)/L10NSharp.dll.
WeSay does not (yet) use this reference itself, but the DLL must be in
the output directory to allow Chorus to work. This also means some tests
fail if we don't have it.

TeamCity has been configured to obtain this dependency. Developers will
for now have to manually add it to lib/{Release/Debug} as appropriate.