Correct WeSay hang on exit after USB enumeration
WeSay hung on exit after the automatic backup or if the Send/Receive
button was pressed.  This resulted in chorus calling palaso to
get an enumeration of the USB drives present.  This in turn caused
NDesk to keep a thread open until the interface is closed.  The
change here is to close system Bus in NDesk.  Expect to modify
this soon to provide some interface within palaso or chorus that
will shut this down instead of doing it from this level.

Change-Id: Icafd84772309f41479a15023ad50430e5ae1bf72