Fix Export to PDF Dictionary issues
Fixed two problems.  If the project name contained spaces,
the export failed.  This was fixed by putting quotes around
the file name for the PDF export.  The second problem was
that when using oriental fonts, the conversion from Odt file
to PDF file displayed the oriental fonts as boxes.  Got
online and saw another person with a similar problem with
Japanese with the recommendation to use invisible instead
of headless.  This seems to work fine and produces the
correct output.

Change-Id: Id5f807142004de04dac79adb957f083408dedb84