First pass at renaming WeSay Configuration Tool
The tools that generate Linux package dependencies currently don't
like filenames with spaces. Although this is clearly a bug, it's
easier to fix our filenames.

Also, since the name of the .exe isn't user-visible, there's no real
need to make it look conversational.

This commit changes the minimum needed for Linux packaging.
Corresponding changes are needed in other places, especially for the
Windows installer.

Change-Id: I49b97f81d14d658509f810db76e9d503726bad59