update vagrant: .gitattributes, readme, urls
Create .gitattributes so that sh files are checked out with unix
line endings.  These scripts can be checked out on Windows and
then shared to linux and they won't work with Windows line
endings.  Set bat files in case we create Windows VMs.

Remove the vagrant modifications to the main .gitattributes file.

Add documentation for how to use vagrant.

Change the box_url values to be the full urls since the short
urls don't really help that much and this will be easier to check
the file paths.

Change-Id: I58c7eeb7be6954b44680fc4ab090cef0052197f1