Handle libdconf dependency in package build
Whether or not dconf is used depends on the version of dconf: it has
to be at least 0.14. Ubuntu Precise uses dconf 0.12 (package name
libdconf0) but later Ubuntu releases have 0.14 or later (package name
libdconf1).

Dependencies are normally handled automatically, but the debhelper
tools aren't able to detect that our code makes runtime decisions
about whether to use dconf. So we have to provide additional
information to tell them that dconf isn't actually used if it's too
old, even though there's a DllImport of it in the assemblies. This
isn't just a case of avoiding an unnecessary dependency, either: the
dllmap for the assembly uses the library name from libdconf1 and the
package fails to build if that library doesn't exist.

The use of dconf is actually in libpalaso's PalasoUIWindowsForms.dll,
so a similar workaround will be needed in any package that
incorporates libpalaso.

A better solution would be to change libpalaso to use gsettings
instead of dconf, since this is a more stable API.

Change-Id: I333aeb6f7ba192c488a85e63d3687d9c5f957282