Fix Mercurial directory hieracrchy
This is an update to 2c5e675 Files in the base directory of Mercurial
were getting copied into a subdirectory of the same name as the file
(e.g. Mercurial/hg was being copied to Mercurial/hg/hg).

There are other places in build.mono.proj where 
%(RecursiveDir)%(Filename)%(Extension) are used and they are not
exhibiting this problem.  I attempted to remove %(Filename)%(Extension) in
those cases and the build broke.

This may be a bug in xbuild or Copy task implementation.

Change-Id: Ie07cfd9ac5d353e7f87c4bfd2ad6498a1d3107ab