Fix missing check mark for Congratulations message
The reported issue was for missing Icon, but was a Text assigned
"Check Mark" in Arial font. The checkmark was also initialized as a
string in a UTF-16LE BOMed file. This was resolved by changing the
font to MS Sans Serif and initializing as a char.
Removed the BOM since it is not needed in the file any more.

Change-Id: I8e3b0cc1551b7fab936bfeb4285b56efcc794b6c