Correct GeckoBox font size and ListView size
After the last change, the font displayed in the GeckoBox was
the wrong size because I was depending up setting the control
font, which someone reset after I set it.  The current
builds a font from the writing system and then sets the
HTML from that.

The list box size did not work properly on linux.  The current
one adjusts the font size up by one which I think takes care
of rounding. It also leaves one entry for a horizontal scroll.
It seems to size correctly for all but my Chinese
project, where GeckoFX seems to be displaying more entries in
the list box than specified by the HTML.

Also created a bold option, although it is not currently used
by anyone, since I realized that what was in there would
not work.

Change-Id: Id7373fca19d93870d1c61e08825ffcb585c02407