Base Gecko Control boxes off of a base class
One of the comments from the initial Gecko implementation review
was that all of the code should be derived from a base class to
avoid code duplication.  This implements that change.

GeckoListBox.cs is not included yet because it is not used yet
and is undergoing major changes to use it.  When it is checked
in next, it will use the same base class.

Change-Id: I0945603ab777968d434b1016b60cc17a31e13b93