Prevent crash in GatherSemanticDomain
If an item was selected from the list box
in GatherSemanticDomain and multiple entries
matched the lexical form, a crash could result.
This was first of all because the array saving
the entry that matched the form and meaning was
overwritten with each iteration of the loop so
that the sense was not saved unless it was the
last entry processed.  The code crashed because
those accessing the saved list assumed there was
at least one entry populating the list.

The change first of all only instantiates the
list being created the first time through the
loop.  Checks were also added to the crash
location to ensure that it contains an entry
to be processed before it attempts to access
the entries.

Change-Id: I45a63c4c1095b7dae9376e6447ac8b6cce6b160f