Make vagrant-cachier plugin optional
There is a bug in the Windows version so that the apt_lists are saved
with the date of the cache file creation instead of the date they
were fetched from the repo.  So you can't update packages.  I have
submitted an issue to the vagrant-cachier repo.
https://github.com/fgrehm/vagrant-cachier/issues/106

Make the use of vagrant-cachier optional so that that vagrant creation
of VMs on Windows can get updates from the APT repos.

Change-Id: I14f05527852ba60c54df4e8caebd6a4b396ccba6