Implement WeSayListBox for Gecko
Added interface to WeSayListBox for the joint controls.
Modified GeckoListBox to use the GeckoBase and to
correctly implement the new interface.  Modified
the two classes using WeSayListBox to now use the
generic interface.  Modified WeSayWordsProject to
register the correct type of control.

Noticed that a cut/paste error had every gecko class
HTML start with "header" and end with "head".  This
caused major problems with the list box class so
I changed this in the other classes once I discovered
it since it is obviously wrong.

Made a couple of other minor changes to the listview
class that should have been made when the base class
was created.

Change-Id: Idb19dec1cad1f28868698ab72d99e3e32d291cb3