Tool Tips are truncated
WS-54 Send/Receive, CAWL Word List & History
tooltips are truncated.
This only occurs on Linux.  It appears that the
calculation of the height of the text by text
renderer can miscalculate scrolling occasionally
and return a height the is one line off.  The
correction made here computes the tooltip length
as if the description was one line longer.  The
tooltip box is then created large enough to
accomodate all of the text.

Change-Id: I4cce5dfb2328fbbfca05675dba65f1a01d4831ea