WS-62 Text entry area does not render properly
On Linux systems, with specific fonts such as Arial 12pt,
entry into a text box displays with the characters
oddly compressed and occasionally invisible.  This change
corrects the display of those characters.

The JIRA also reports issues with the cursor position.
Once displayed properly, the issue is somewhat less
apparent, but can still be off by one character.  This
change does not address that part of the problem.
Changing the font by one pt can be a workaround.  Arial
11pt and Arial 13pt do not display the same issues and
the same seems to be true for the one other font I saw.

Change-Id: I411399267c00bb7ccce69915753d8c3caea65dc2