WS-63 CAWL-MOZ Entries do not display in list box
Except for the first two entries, selecting an entry
and then providing a word does not add the word to
the list box and does not add the meaning to the
word when it is added to the dictionary.

The problem was that, except for the first two entries,
English definitions were not defined for the word.
Also Portugese forms were not provided.  This was
a customer provided feature that was not completely
validated.  I adjusted the lift file to conform to the
expectations of the code for the CAWL and CAWL-MOZ lift
files.

Change-Id: Ia315ca6ea98871bc104d77a5da870dba6ba2ad4b