WS-66 Adding note to a word in DBE fails
If you have a project name such as Espanol (with
a tilda n), the tilda was being converted to
UTF-8 %C3%B1.  When chorus tries to create a
note later using the modified name, mono
Uri.cs returns a failure because it validates
that each character is a number or a character.
Percent doesn't qualify.  Tilda n is recognized
as a character as are some other special
characters I checked.   This also fails in
Windows.  The line in question only appears to
be used as part of opening chorus notes, and
since it creates the failure, I'm removing it.

Change-Id: I803acaf4ea1d1f922259caa0b24aac4d812c7457