Fixed some Gecko browser mode issues
Corrected spelling of longestMeaningWord.
Fixed problem with ComboBox on semantic domain control
being clipped in some instances.
Fixed problem with blank sometimes appearing in the
word entry of the gather by semantic domain when
you click on another word in the list.
Fixed inadvertent change to the font of the
gather by semantic domain combo box when I moved
the setting of some WritingSystems up as part of
an earlier change but failed to move the default.
This resulted in the default case always being
used.

Change-Id: I09e899c99599ed5c30e909876329523b2b688ab1