Various gecko mode fixes
1) Change list box and combo box background color from
the default light yellow to white.
2) In Dictionary Browse and Control and the Missing Items
tabs, change so that in Gecko mode, the splitter window can't
stay smaller than the width of the list box.  It adjusts on
initialization and checks each time words are selected
3) Gecko 29 in Windows doesn't display properly.  A workaround
for this problem is to minimize and maximize the window as
the last part of setup.  This is only done for Windows when
in Gecko Mode.
4) Noticed that some of the properties were set after the
setting of the data source, which is the point where they
would be used initially.  Moved them to before the setting
of the data source.
5) Improve list view functionality to recompute the number of
items that can be viewed.  It usually works correctly, but
can be a problem if the default word system for the word
field is empty and an alternate form is used instead.

Change-Id: I6822e967967be829bbacbe80a2e3be85f9248e59