Add ENV variable to control XulRunner Init
There still seems to be issues with WeSay crashing when XulRunner
is initialized.  It seems to be that C# exceptions that would
normally be caught by the C# code are instead causing SIGSEGV
faults and crashing the app.  Use WESAY_INIT_XULRUNNER=0
to disable initialization of XulRunner (still enabled by default).

Change-Id: I54dda3443e13ccd34b6ed336fe680c9e59d88cf8