WS-59 Word list not repainted correctly
Word list in "Dictionary Browse and Edit" does
not repaint properly.   With some long entries,
if the scroll bar on the list view is pushed
all the way to the right and then back to the
left, text does not display properly.  Can be
cut off and distorted.

Fix was to invalidate on horizontal scroll.
Problem does not appear in Windows.

Change-Id: I2ec3e73be8709b52229358360e52cf3d899cf320