WS-94 Crash on "Part of Speech" tab
Completing an item, clicking on the last
uncompleted item, and then clicking on an
entry in the completed item crashes. The
FocusIndex for Mono ends up being out of
scope once the item is removed after
clicking on the last entry and removing
the current item.  SelectedIndex adjusts
as you would expect, but not the focus.
Works fine in Windows or if running
Gecko option without change.

Change-Id: I9f60efa901b19e58e9c15ede514832d4cd02675f