WS-95 Hang on add of picture
This change is to supplement an upcoming palaso change that
will correct the problem.  The system hangs on import of a
.png file.  The call to exiftool hangs in palaso.
The palaso fix will correct that, but introduces
a new failure because WeSay gives all file names an extension
of .png.  This change preserves the extension of the name of
the picture file being imported and allows WeSay to run with
the new fix for palaso when that is released.  It also works
with the current palaso, although png files will still hang.

Change-Id: Ib22c27e56f54815616126b5630c19e6533c524b7