Build with icu48 or icu52 depending on icu-config
For Trusty, ICU was updated to version 52.  In Precise and Saucy,
ICU is still at version 48. The way that icu.net.dll is built,
there needed to be separate builds to support each of these
versions.  So change the wesay build to download both artifacts
and copy the one during a build to lib/$(Configuration) depending
on the ICU version (using icu-config --version).

Change-Id: I198bde4872340b85f22510c0326cf704f5fec7f2