Changes to improve Gecko text and list box
Introduce a delay before changing the focus to the
inner element when the browser receives focus.
When the element is clicked on, reset the focus and
the current text contents.
Allow multitextcontrol to select WeSayTextBox and
GeckoBoxes instead of just WeSayTextBox.
Change list box to keep selected item visible.
Change list box to leave the background transparent.

Change-Id: I2facf77bfcd81d851f7b5e41275189b9ca0c27fa