Correct Collapsible Splitter error
In some cases, Mono will return a -1 for the split
position instead of the current position.  This
was causing WeSay to hide the control.  This change
causes WeSay to neither save nor adjust the control
if it receives a value that is invalid.

Also set up to retry the adjustment of the splitter
one time after an initial failure of this kind.  This
corrected both of the problems the tester reported.

Change-Id: Ia1d5f4ea4ed1e850931d2407264cf1b95a44df66