Add Chorus_Help.chm to Linux package
The Windows install already includes Chorus_Help.chm in the
ChorusMergeModule.msm.  The Linux package building didn't include
it.  This downloads the file from a TeamCity build and has it
installed into the InstallApp directory (instead of InstallData)
since this is a Chorus requirement and JasonNaylor told me that
I should not change that behavior.

Change-Id: Iece7c1e3140bc507429698160a5c481549f938a5