WS-105 PageDown in CAWL-MOZ display issue
If the last word in the list is selected and then
the page down key is pressed, the "Congratulations"
display appears, overlaying most of the screen.  In
Windows, the edge of the list box isn't visible, but
in Linux, it is.  This change moves the Congratulations
control to the left so  that it overlays all of the
list box in Linux and Windows and doesn't provide
confusion.

Change-Id: I29b32f07e36ff114e8e3b35cec400f833cf9f70f