WS-152 Error box when adding pictures
WeSay determines the name the picture being added is saved
as from the meaning entry.  If it contains illegal characters,
the save ultimately fails.  This change replaces any illegal
characters with an underscore character. If this should result
in a dupicate to an existing file name, the code will go
into the existing logic and resolve it by attaching a
timestamp to create the new name.

Change-Id: I31dca4902bd894736dc98e95c94b5cfbb59c9a7a