WS-156 Turning on Gecko causes error
This corrects problems that were present when running
on Windows with Gecko turned on.  Some of the routines
recently added for fixing Linux problems should not be run.
In addition, there were some issues with tabbing out of
fields on Windows not updating the entry properly.  This
fixes both issues.

Change-Id: Ic028a84eb0b44b7610c4d6b1937e8822a6603fac