Use web-friendly format for images (like in 1.4)
In WS-34711, there were changes to make a copy of the Image before
saving. Also, there were changes to use a web-friendly format for
images (either Png or Jpeg). However, there were bugs in how this
was implemented. The extension (in WeSay) was always set to png,
but the format of the save files was determined by the format of
the imported image. So the file was stored in either Png or Jpeg
format, but always had a .png extension. In WS-152, the code was
fixed to maintain the file extension of the original imported file,
however the format was still being saved as either Png or Jpeg.

PalasoImage.Save was changed to have the format match the file
extension of the file being saved.
PalasoImage.FileExtensionForWebFormat was added to all the caller
to save the image in a web-friendly format.

Change-Id: I1962cbb70a7b91d011049fc4f34c99643958836e