WS-161 SE Asian scripts are not correct
Additional changes were required to deal with
cases where the language needed to be specified
to identify the font type that should be used
where the font supported multiple languages. In
particular, this made the khamti language show
the correct format instead of the Burmese format
supported by the Padauk font by default.

Change-Id: I3d4c78f5a1da3c961a96443b97d2f4b1a4302fe1