WS-211 Show pictures in OpenOffice Documents
The <illustration> element in LIFT has a href with the relative
path to the picture.  On Linux, the path separator is forward
slash, however on Windows it is backslash (bug?). The paths in
OpenOffice documents have to be forward slashes.

The initial implementation of the OpenOffice document addin was
developed on Linux so these matched.  The transform now accounts
for either.

Change-Id: Iaa5d3454900bdfdd59b46e6ac540d3d795e0afc4