WS-213 List box appears in DB&E
List box from Base Forms tab does not disappear
if the "+" key is pressed to add the word.  If
you then click the Dictionary Browse and Edit tab,
the list box is still present.

This also addresses a crash that was seen when
exiting the program from the Base Forms tab.  It
used a control that had already been disposed.

Change-Id: Iefce4eb4ac8870ebd3926531082e627c75605a2d