WS-239 Horizontal and vertical scroll bars appear in Gecko
Horizontal and vertical scroll bars would appear if
a long word entry was made without spaces in it.  If
spaces were entered, the entry wrapped at the spaces
otherwise it would not break and the scroll bars
appeared instead.  With the change, it still works
the same for entries with spaces but it will now also
wrap at the end of the long entry.

Change-Id: I8604e4eae96e8c15a1a95a87d85acf17114e7376