WS-248 Semantic domain localization issues
Changed the field containing the question to a
IWeSayTextBox field.  The field is set as a
non editable multiline field.  Also corrected
the Gecko combo box so that the vertical scroll bar
never appears.  With the first change, the question
now always appears in full on the screen.  It is
not truncated in Windows or Linux.

Change-Id: I321a761db0069f7a293bad9ae964bf04f045bc65