WS-244: Fix crash in focus change with ListViews
When a user deletes an entry in the Database Browse & Edit tab,
the code selects another entry to be the SelectedIndex.  When
the ListView receives FocusChanged, it invalidates the current
focused item.

If the last item in a list is removed, the selection to the
previous item was changed, but not the focus.  So ListView
would crash when trying to invalidate one past the last item.

Since WeSayListView is changing how Focus and Selection is
handled with a standard ListView, it should make sure that
the focus is set to the SelectedIndex when that changes.

WS-94 was an attempt to fix this in the MissingInfoControl.

Change-Id: I9b9091abe2e7747570e006853f74d37d39950568