WS-280: Fix crash when using MouseWheel
In Windows, you can disable the MouseWheel by handling the
Control.MouseWheel event and casting the MouseEventArgs param
to HandledMouseEventArgs and setting Handled = true.  However,
Mono doesn't implement this.  For now, check to see if it can be
casted safely.  To disable the scroll wheel on Linux will require
fixing Mono.  See:
https://bugzilla.xamarin.com/show_bug.cgi?id=25201

Change-Id: I1392eeace7fa07e9c8bf52218965a98d0049e2ce